Moved tests to subfolders.
