+++ src/Tests/Tests.Logic.Core/Tests.Logic.Core.csproj
@@ -56,28 +56,28 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BaseDataModelTests.cs" />
-    <Compile Include="BaseObservableObjectTests.cs" />
-    <Compile Include="BaseTest.cs" />
+    <Compile Include="Tests\BaseClasses\BaseDataModelTests.cs" />
+    <Compile Include="Tests\BaseClasses\BaseObservableObjectTests.cs" />
+    <Compile Include="Tests\BaseClasses\BaseTest.cs" />
     <Compile Include="Extensions\TestDataModelExtensions.cs" />
-    <Compile Include="MessengerTests.cs" />
+    <Compile Include="Tests\MessengerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RangeAttributeCheckerTests.cs" />
-    <Compile Include="RegularExpressionAttributeCheckerTests.cs" />
-    <Compile Include="RequiredAttributeCheckerTests.cs" />
-    <Compile Include="MinLengthAttributeCheckerTests.cs" />
+    <Compile Include="Tests\AttributeCheckers\RangeAttributeCheckerTests.cs" />
+    <Compile Include="Tests\AttributeCheckers\RegularExpressionAttributeCheckerTests.cs" />
+    <Compile Include="Tests\AttributeCheckers\RequiredAttributeCheckerTests.cs" />
+    <Compile Include="Tests\AttributeCheckers\MinLengthAttributeCheckerTests.cs" />
     <Compile Include="TestResources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>TestResources.resx</DependentUpon>
     </Compile>
-    <Compile Include="StringLengthAttributeCheckerTests.cs" />
+    <Compile Include="Tests\AttributeCheckers\StringLengthAttributeCheckerTests.cs" />
     <Compile Include="TestHandler.cs" />
     <Compile Include="TestModels\LazyTestDataModel.cs" />
     <Compile Include="TestModels\TestDataModel.cs" />
     <Compile Include="TestModels\TestObservable.cs" />
     <Compile Include="Variables.cs" />
-    <Compile Include="ValidationAttributeCheckResultTests.cs" />
+    <Compile Include="Tests\ValidationAttributeCheckResultTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="CodeCoverage.runsettings" />
