+++ src/Tests/Tests.Logic.Core/Tests/AttributeCheckers/RangeAttributeCheckerTests.cs
@@ -1,14 +1,13 @@
-﻿namespace deveer.DoctorFlox.Tests.Logic.Core
+﻿namespace deveer.DoctorFlox.Tests.Logic.Core.Tests.AttributeCheckers
 {
     using System;
-    using System.CodeDom;
-    using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.Diagnostics.CodeAnalysis;
     using System.Linq;
 
+    using BaseClasses;
+
     using devdeer.DoctorFlox.Helpers.Validation;
-    using devdeer.DoctorFlox.Models;
 
     using Extensions;
 
@@ -26,7 +25,7 @@
         #region methods
 
         /// <summary>
-        /// Checks if the <see cref="RangeAttributeChecker.InternalCheck"/> method reacts as expected.
+        /// Checks if the <see cref="RangeAttributeChecker.InternalCheck" /> method reacts as expected.
         /// </summary>
         [TestMethod]
         [DeploymentItem(@"DataSources\RangeAttributeCheckerInternalCheckTest.csv")]
@@ -35,19 +34,19 @@
         {
             // arrange
             var expectedValid = GetConvertedTestValueAsBool(0) ?? throw new ApplicationException("Could not retrieve expected value.");
-            var expectedErrorMessageKey  = GetConvertedTestValueAsString(1);
+            var expectedErrorMessageKey = GetConvertedTestValueAsString(1);
             var expectedErrorMessage = string.IsNullOrEmpty(expectedErrorMessageKey) ? string.Empty : TestResources.ResourceManager.GetString(expectedErrorMessageKey);
             var item = new TestDataModel
             {
-               Salary = GetConvertedTestValueAsInt(2) ?? 0
+                Salary = GetConvertedTestValueAsInt(2) ?? 0
             };
             var checker = new RangeAttributeChecker();
-            var keyAttribute = item.GetAttribute<RangeAttribute>(nameof(TestDataModel.Salary));            
+            var keyAttribute = item.GetAttribute<RangeAttribute>(nameof(TestDataModel.Salary));
             // act 
-            var result = checker.Check(keyAttribute, item.Salary);            
+            var result = checker.Check(keyAttribute, item.Salary);
             // assert
             Assert.AreEqual(expectedValid, result.IsValid, "The validity check for salary failed.");
-            Assert.AreEqual(expectedErrorMessage, result.ErrorMessage, "The error message for salary is unexpected.");            
+            Assert.AreEqual(expectedErrorMessage, result.ErrorMessage, "The error message for salary is unexpected.");
         }
 
         #endregion
