+++ src/Tests/Tests.Logic.Core/Tests/AttributeCheckers/RegularExpressionAttributeCheckerTests.cs
@@ -1,14 +1,13 @@
-﻿namespace deveer.DoctorFlox.Tests.Logic.Core
+﻿namespace deveer.DoctorFlox.Tests.Logic.Core.Tests.AttributeCheckers
 {
     using System;
-    using System.CodeDom;
-    using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.Diagnostics.CodeAnalysis;
     using System.Linq;
 
+    using BaseClasses;
+
     using devdeer.DoctorFlox.Helpers.Validation;
-    using devdeer.DoctorFlox.Models;
 
     using Extensions;
 
@@ -26,28 +25,32 @@
         #region methods
 
         /// <summary>
-        /// Checks if the <see cref="RegularExpressionAttributeChecker.InternalCheck"/> method reacts as expected.
+        /// Checks if the <see cref="RegularExpressionAttributeChecker.InternalCheck" /> method reacts as expected.
         /// </summary>
         [TestMethod]
         [DeploymentItem(@"DataSources\RegularExpressionAttributeCheckerInternalCheckTest.csv")]
-        [DataSource(@"Microsoft.VisualStudio.TestTools.DataSource.CSV", "RegularExpressionAttributeCheckerInternalCheckTest", "RegularExpressionAttributeCheckerInternalCheckTest#csv", DataAccessMethod.Sequential)]
+        [DataSource(
+            @"Microsoft.VisualStudio.TestTools.DataSource.CSV",
+            "RegularExpressionAttributeCheckerInternalCheckTest",
+            "RegularExpressionAttributeCheckerInternalCheckTest#csv",
+            DataAccessMethod.Sequential)]
         public void RegularExpressionAttributeCheckerInternalCheckTest()
         {
             // arrange
             var expectedValid = GetConvertedTestValueAsBool(0) ?? throw new ApplicationException("Could not retrieve expected value.");
-            var expectedErrorMessageKey  = GetConvertedTestValueAsString(1);
+            var expectedErrorMessageKey = GetConvertedTestValueAsString(1);
             var expectedErrorMessage = string.IsNullOrEmpty(expectedErrorMessageKey) ? string.Empty : TestResources.ResourceManager.GetString(expectedErrorMessageKey);
             var item = new TestDataModel
             {
-               Key = GetConvertedTestValueAsString(2)
+                Key = GetConvertedTestValueAsString(2)
             };
             var checker = new RegularExpressionAttributeChecker();
-            var keyAttribute = item.GetAttribute<RegularExpressionAttribute>(nameof(TestDataModel.Key));            
+            var keyAttribute = item.GetAttribute<RegularExpressionAttribute>(nameof(TestDataModel.Key));
             // act 
-            var result = checker.Check(keyAttribute, item.Key);            
+            var result = checker.Check(keyAttribute, item.Key);
             // assert
             Assert.AreEqual(expectedValid, result.IsValid, "The validity check for key failed.");
-            Assert.AreEqual(expectedErrorMessage, result.ErrorMessage, "The error message for key is unexpected.");            
+            Assert.AreEqual(expectedErrorMessage, result.ErrorMessage, "The error message for key is unexpected.");
         }
 
         #endregion
